+++ crates/sncast/src/response/account/create.rs
@@ -1,5 +1,5 @@
 use conversions::padded_felt::PaddedFelt;
-use foundry_ui::{Message, formats::OutputFormat};
+use foundry_ui::{Message, OutputFormat};
 use serde::{Serialize, Serializer};
 use starknet_types_core::felt::Felt;
