+++ crates/sncast/src/response/cast_message.rs
@@ -1,4 +1,6 @@
-use foundry_ui::{Message, formats::NumbersFormat};
+use foundry_ui::Message;
+
+use crate::NumbersFormat;
 
 pub struct CastMessage<T: Message> {
     pub numbers_format: NumbersFormat,
