+++ crates/sncast/src/response/show_config.rs
@@ -1,5 +1,5 @@
 use camino::Utf8PathBuf;
-use foundry_ui::{Message, formats::OutputFormat};
+use foundry_ui::{Message, OutputFormat};
 use serde::Serialize;
 
 use crate::helpers::block_explorer;
