+++ crates/sncast/src/response/tx_status.rs
@@ -1,5 +1,5 @@
 use conversions::serde::serialize::CairoSerialize;
-use foundry_ui::{Message, formats::OutputFormat};
+use foundry_ui::{Message, OutputFormat};
 use serde::Serialize;
 
 use super::{
