+++ crates/sncast/src/starknet_commands/account/list.rs
@@ -3,12 +3,12 @@ use camino::Utf8PathBuf;
 use clap::Args;
 use conversions::string::IntoDecStr;
 use conversions::string::IntoHexStr;
-use foundry_ui::formats::NumbersFormat;
-use foundry_ui::formats::OutputFormat;
+use foundry_ui::OutputFormat;
 use itertools::Itertools;
 use serde::Deserialize;
 use serde::Serialize;
 use sncast::AccountType;
+use sncast::NumbersFormat;
 use sncast::{AccountData, NestedMap, check_account_file_exists, read_and_parse_json_file};
 use std::collections::HashMap;
 use std::fmt::Display;
